Fixes the container __call methody.
The __call method now calls the parent DI Container __call method
in the event that the requested item exists in the container.
If it does not exist then it runs the router code we had before.
Unfortantly we still can't use __callStatic, that would be the
most elegant solution.

This also resolves #1.
codixor you can now grab the request object from with in the route.